implemented but needing refactor
