+++ lib/components/Admin.js
@@ -44,25 +44,6 @@ export default class Admin extends Component {
     })
   }
 
-  assignRooms(spikes) {
-    let classRooms = { a: '', b: '', c: '', d: '', e: '', f: '', g: '', h: '', auditorium: '', vault: '' }
-    spikes.forEach((spike) => {
-      if(spike.location) {
-        if (spike.location === 'Classroom A') classRooms.a = spike.title;
-        if (spike.location === 'Classroom B') classRooms.b = spike.title;
-        if (spike.location === 'Classroom C') classRooms.c = spike.title;
-        if (spike.location === 'Classroom D') classRooms.d = spike.title;
-        if (spike.location === 'Classroom E') classRooms.e = spike.title;
-        if (spike.location === 'Classroom F') classRooms.f = spike.title;
-        if (spike.location === 'Classroom G') classRooms.g = spike.title;
-        if (spike.location === 'Classroom H') classRooms.h = spike.title;
-        if (spike.location === 'Grand Room') classRooms.auditorium = spike.title;
-        if (spike.location === 'Vault Study') classRooms.vault = spike.title;
-      }
-    })
-    return classRooms;
-  }
-
   render() {
     const {
     	admins,
@@ -71,7 +52,8 @@ export default class Admin extends Component {
     	removeAdmin,
     	spikes,
     	updateSpike,
-    	user
+    	user,
+      assignRooms
     } = this.props;
 
 
@@ -101,7 +83,7 @@ export default class Admin extends Component {
         newAdmin={newAdmin}
       />
 
-    let roomAssignments = this.assignRooms(spikes)
+    let roomAssignments = assignRooms(spikes)
 
     return (
       <section className='Admin'>
