+++ lib/components/App.js
@@ -162,6 +162,25 @@ export default class App extends Component {
 			.replace(/<![\s\S]*?--[ \t\n\r]*>/gi, '');
   }
 
+  assignRooms(spikes) {
+    let classRooms = { a: '', b: '', c: '', d: '', e: '', f: '', g: '', h: '', auditorium: '', vault: '' }
+    spikes.forEach((spike) => {
+      if(spike.location) {
+        if (spike.location === 'Classroom A') classRooms.a = spike.title;
+        if (spike.location === 'Classroom B') classRooms.b = spike.title;
+        if (spike.location === 'Classroom C') classRooms.c = spike.title;
+        if (spike.location === 'Classroom D') classRooms.d = spike.title;
+        if (spike.location === 'Classroom E') classRooms.e = spike.title;
+        if (spike.location === 'Classroom F') classRooms.f = spike.title;
+        if (spike.location === 'Classroom G') classRooms.g = spike.title;
+        if (spike.location === 'Classroom H') classRooms.h = spike.title;
+        if (spike.location === 'Grand Room') classRooms.auditorium = spike.title;
+        if (spike.location === 'Vault Study') classRooms.vault = spike.title;
+      }
+    });
+    return classRooms;
+  }
+
   render() {
     const user = this.state.user;
     if (!user) { return <SignIn /> }
@@ -184,6 +203,7 @@ export default class App extends Component {
             removeAdmin={(index) => this.removeAdmin(index)}
             user={this.state.user}
             speaker={this.state.speaker}
+            assignRooms={this.assignRooms}
           />
         </section>
       )
@@ -208,6 +228,7 @@ export default class App extends Component {
             spikes={this.state.spikes}
             user={user}
             updateCount={(spike) => this.updateCount(spike)}
+            assignRooms={this.assignRooms}
           />
         </section>
       )
