+++ lib/components/Spikes.js
@@ -2,15 +2,17 @@ import React , { Component } from 'react';
 import firebase from '../firebase';
 import { map, uniq } from 'lodash';
 import moment from 'moment';
+import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
 
-
+import FloorMap from './FloorMap';
 import Spike from './Spike';
 
 export default class Spikes extends Component {
   constructor(props) {
     super();
     this.state = {
-      dateFilter: 'all'
+      dateFilter: 'all',
+      floorMap: false,
     };
   }
 
@@ -46,14 +48,29 @@ export default class Spikes extends Component {
   }
 
   render() {
-    const { spikes } = this.props
+    const { spikes, assignRooms } = this.props
 
     let allSpikes = this.filterSpikes(spikes)
 
     let allDateOptions = this.dateOptions(spikes)
 
+    let roomAssignments = assignRooms(spikes);
+
     return (
       <section>
+        <button
+          className='ShowMapButton'
+          onClick={() => this.setState({ floorMap: true })}
+        >
+          Show Map
+        </button>
+        <ReactCSSTransitionGroup
+          transitionName='AnimateForm'
+          transitionEnterTimeout={700}
+          transitionLeaveTimeout={700}
+        >
+          {this.state.floorMap && <FloorMap rooms={roomAssignments} hideMap={()=>this.setState({ floorMap: false })}/>}
+        </ReactCSSTransitionGroup>
         <select name="all" onChange={(e)=>this.setState({ dateFilter: parseInt(e.target.value) })
         }>
           <option value={Date.now()}>Current Week</option>
